src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aniso8601==7.0.0          # via graphene
certifi==2020.6.20        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
coverage==4.5.4           # via -r requirements.in
cryptography==3.2         # via jwcrypto
django-cors-headers==3.5.0  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-graphql-jwt==0.3.1  # via -r requirements.in
django-jenkins==0.110.0   # via -r requirements.in
django==2.2.16            # via -r requirements.in, django-cors-headers, django-filter, django-graphql-jwt, django-jenkins, graphene-django
flake8==3.8.4             # via -r requirements.in
graphene-django==2.13.0   # via -r requirements.in, django-graphql-jwt
graphene==2.1.8           # via graphene-django
graphql-core==2.3.2       # via django-graphql-jwt, graphene, graphene-django, graphql-relay
graphql-relay==2.0.1      # via graphene
idna==2.10                # via requests
jwcrypto==0.8             # via python-jwt
mccabe==0.6.1             # via flake8
pep8==1.7.1               # via -r requirements.in
pip-tools==5.3.1          # via -r requirements.in
promise==2.3              # via graphene-django, graphql-core, graphql-relay
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via -r requirements.in, flake8
pyjwt==1.7.1              # via django-graphql-jwt
python-jwt==3.3.0         # via -r requirements.in
pytz==2020.1              # via django
requests==2.24.0          # via -r requirements.in
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.15.0               # via cryptography, graphene, graphene-django, graphql-core, graphql-relay, pip-tools, promise, singledispatch
sqlparse==0.4.1           # via django
unidecode==1.1.1          # via graphene-django
urllib3==1.25.11          # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
